+++ Assets/Resources/Materials/DefaultMat.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 66c79608eaae6ca42b5fb7ff459e65f4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
