+++ Assets/Resources/Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 85cc1e08b09ee0e4c91a88038b3e10c4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
