+++ Assets/Resources/Textures/grass.kra.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ff1414f8e72a676439c60782571ade34
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
