+++ Assets/Scripts/ChunkScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 31d0c852e86bb6548b1e187d1d8b2ffd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
