+++ Assets/Scripts/Utils.cs
@@ -1,16 +1,21 @@
 using UnityEngine;
 using System.IO;
+
 #nullable enable
+
 class Utils
 {
-    public static GameObject LoadPrefabFromFile(string filename)
+  public static Object LoadPrefabFromFile(string filename)
+  {
+    Debug.Log("Trying to load LevelPrefab from file (" + filename + ")...");
+
+    var loadedObject = Resources.Load(filename);
+
+    if (loadedObject == null)
     {
-        Debug.Log("Trying to load LevelPrefab from file (" + filename + ")...");
-        var loadedObject = Resources.Load(filename);
-        if (loadedObject == null)
-        {
-            throw new FileNotFoundException("...no file found - please check the configuration");
-        }
-        return (GameObject)loadedObject;
+      throw new FileNotFoundException("...no file found - please check the configuration");
     }
+
+    return loadedObject;
+  }
 }
